backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slate-backend"
version = "0.1.0"
description = "Slate Python backend powered by FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "pydantic-settings>=2.7,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "PyJWT>=2.10,<3.0",
  "redis>=5.2,<6.0",
  "SQLAlchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.9,<1.0",
]

[tool.setuptools.packages.find]
include = ["slate_api*"]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
